Update index.js
Changed the code so that the text within endorsements uses the existing CSS selector for the paragraphs. I also did the same for the white endorsement boxes and their CSS selector.
Was trying to change the code so that any text typed in the input-field is shown at the top-left of the input-field instead of the center